sc => bf (part of showcase => BitFlux rename)
