+++ src/assets/js/model/menu/productAdaptor.js
@@ -1,4 +1,3 @@
-// Generates a menu option similar to those generated by sc.model.menu.option from a sc.model.data.product object
 export default function(product) {
     return {
         displayString: product.display,
